requirements.txt
fastapi==0.95.1
motor==3.1.2
pandas==2.0.0
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
scikit_learn==1.2.2
uvicorn==0.21.1